Switch to wasi for native
The combination of wasm-bindgen and getrandom (which is unavoidable due to bevy relying on ahash which relies on getrandom) was causing linking function issues that I wasn't able work around. Admittedly, I'm very green when it comes to rust and to wasm, but I was able to bypass the issues by implementing changes that @TurboSnailGit had made in their fork.

The magic sauce was:
    let mut linker: Linker<WasiModState> = Linker::new(engine);
    wasi_common::sync::add_to_linker(&mut linker, |s| &mut s.wasi_ctx)?;

which solved my import woes with wasm-bindgen and getrandom since it would handle all of the importing of definitions